breaking down a large method into smaller ones

+++ source/ClassElectronicInvoiceUserInterface.php
@@ -32,58 +32,11 @@ class ClassElectronicInvoiceUserInterface
     private function actionAnalyzeZIPfromANAFfromLocalFolder(string $strFilePath): array
     {
         $arrayFiles    = new \RecursiveDirectoryIterator($strFilePath, \FilesystemIterator::SKIP_DOTS);
-        $classZip      = new \ZipArchive();
         $arrayInvoices = [];
         $intFileNo     = 0;
         foreach ($arrayFiles as $strFile) {
             if ($strFile->isFile() && ($strFile->getExtension() === 'zip')) {
-                $arrayInvoices[$intFileNo] = $this->setStandardizedFeedbackArray([
-                    'ResponseIndex' => $strFile->getBasename('.zip'),
-                    'Type'          => 'Default',
-                ]);
-                $res                       = $classZip->open($strFile->getRealPath(), \ZipArchive::RDONLY);
-                if ($res === true) {
-                    $intFilesArchived = $classZip->numFiles;
-                    for ($intArchivedFile = 0; $intArchivedFile < $intFilesArchived; $intArchivedFile++) {
-                        $strArchivedFile = $classZip->getNameIndex($intArchivedFile);
-                        $matches         = [];
-                        preg_match('/^[0-9]{5,20}\.xml$/', $strArchivedFile, $matches, PREG_OFFSET_CAPTURE);
-                        $matches2        = [];
-                        preg_match('/^semnatura_[0-9]{5,20}\.xml$/', $strArchivedFile, $matches2, PREG_OFFSET_CAPTURE);
-                        if ($matches !== []) {
-                            $resInvoice        = $classZip->getStream($strArchivedFile);
-                            $strInvoiceContent = stream_get_contents($resInvoice);
-                            fclose($resInvoice);
-                            $intInvoiceSize    = strlen($strInvoiceContent);
-                            if ($intInvoiceSize > 1000) {
-                                $arrayInvoices[$intFileNo] = $this->setStandardizedFeedbackArray([
-                                    'ResponseIndex'       => $strFile->getBasename('.zip'),
-                                    'Size'                => $intInvoiceSize,
-                                    'Type'                => 'XmlFile_Large',
-                                    'LoadingIndex'        => substr($matches[0][0], 0, -4),
-                                    'strArchivedFileName' => $strArchivedFile,
-                                    'strInvoiceContent'   => $strInvoiceContent,
-                                ]);
-                            } else {
-                                $arrayInvoices[$intFileNo] = $this->setStandardizedFeedbackArray([
-                                    'ResponseIndex'     => $strFile->getBasename('.zip'),
-                                    'Size'              => $intInvoiceSize,
-                                    'Type'              => 'XmlFile_Small',
-                                    'strInvoiceContent' => $strInvoiceContent,
-                                ]);
-                            }
-                        } elseif ($matches2 === []) {
-                            echo vsprintf('<div>' . $this->arrayConfiguration['Feedback']['DifferentFile'] . '</div>', [
-                                $strArchivedFile,
-                                $strFile->getBasename(),
-                            ]);
-                        }
-                    }
-                } else {
-                    // @codeCoverageIgnoreStart
-                    throw new \RuntimeException(sprintf('Archive %s could not be opened!', $strFile->getRealPath()));
-                    // @codeCoverageIgnoreEnd
-                }
+                $arrayInvoices[$intFileNo] = $this->setArchiveFromAnaf($strFile->getRealPath());
                 $intFileNo++;
             }
         }
@@ -106,6 +59,45 @@ class ClassElectronicInvoiceUserInterface
             implode(';', $arrayStylePieces),
             $arrayButtonFeatures['Text'],
         ]);
+    }
+
+    private function setArchiveFromAnaf(string $strFile)
+    {
+        $classZip      = new \ZipArchive();
+        $arrayToReturn = [];
+        $res           = $classZip->open($strFile, \ZipArchive::RDONLY);
+        if ($res === true) {
+            $intFilesArchived = $classZip->numFiles;
+            for ($intArchivedFile = 0; $intArchivedFile < $intFilesArchived; $intArchivedFile++) {
+                $strArchivedFile = $classZip->getNameIndex($intArchivedFile);
+                $matches         = [];
+                preg_match('/^[0-9]{5,20}\.xml$/', $strArchivedFile, $matches, PREG_OFFSET_CAPTURE);
+                $matches2        = [];
+                preg_match('/^semnatura_[0-9]{5,20}\.xml$/', $strArchivedFile, $matches2, PREG_OFFSET_CAPTURE);
+                if ($matches !== []) {
+                    $resInvoice        = $classZip->getStream($strArchivedFile);
+                    $strInvoiceContent = stream_get_contents($resInvoice);
+                    fclose($resInvoice);
+                    $arrayToReturn     = $this->setStandardizedFeedbackArray([
+                        'ResponseIndex'       => pathinfo($strFile)['filename'],
+                        'Size'                => strlen($strInvoiceContent),
+                        'Matches'             => $matches,
+                        'strArchivedFileName' => $strArchivedFile,
+                        'strInvoiceContent'   => $strInvoiceContent,
+                    ]);
+                } elseif ($matches2 === []) {
+                    echo vsprintf('<div>' . $this->arrayConfiguration['Feedback']['DifferentFile'] . '</div>', [
+                        $strArchivedFile,
+                        $strFile->getBasename(),
+                    ]);
+                }
+            }
+        } else {
+            // @codeCoverageIgnoreStart
+            throw new \RuntimeException(sprintf('Archive %s could not be opened!', $strFile));
+            // @codeCoverageIgnoreEnd
+        }
+        return $arrayToReturn;
     }
 
     public function setActionToDo(): void
@@ -191,42 +183,39 @@ class ClassElectronicInvoiceUserInterface
             'Customer_Name' => '',
             'Error'         => '',
         ];
-        switch ($arrayData['Type']) {
-            case 'XmlFile_Large':
-                file_put_contents($arrayData['strArchivedFileName'], $arrayData['strInvoiceContent']);
-                $appR                           = new \danielgp\efactura\ClassElectronicInvoiceRead();
-                $arrayElectronicInv             = $appR->readElectronicInvoice($arrayData['strArchivedFileName']);
-                $arrayBasic                     = $arrayElectronicInv['Header']['CommonBasicComponents-2'];
-                $arrayAggregate                 = $arrayElectronicInv['Header']['CommonAggregateComponents-2'];
-                $floatAmounts                   = [
-                    'wo_VAT' => (float) $arrayAggregate['LegalMonetaryTotal']['TaxExclusiveAmount']['value'],
-                    'TOTAL'  => (float) $arrayAggregate['LegalMonetaryTotal']['TaxInclusiveAmount']['value'],
-                ];
-                $arrayParties                   = [
-                    'Customer' => $arrayAggregate['AccountingCustomerParty']['Party'],
-                    'Supplier' => $arrayAggregate['AccountingSupplierParty']['Party'],
-                ];
-                $arrayToReturn['LoadingIndex']  = $arrayData['LoadingIndex'];
-                $arrayToReturn['Size']          = $arrayData['Size'];
-                $arrayToReturn['Document_No']   = $arrayBasic['ID'];
-                $arrayToReturn['Issue_Date']    = $arrayBasic['IssueDate'];
-                $arrayToReturn['Amount_wo_VAT'] = $floatAmounts['wo_VAT'];
-                $arrayToReturn['Amount_TOTAL']  = $floatAmounts['TOTAL'];
-                $arrayToReturn['Amount_VAT']    = ($floatAmounts['TOTAL'] - $floatAmounts['wo_VAT']);
-                $arrayToReturn['Supplier_CUI']  = $arrayParties['Supplier']['PartyTaxScheme']['01']['CompanyID'];
-                $arrayToReturn['Supplier_Name'] = $arrayParties['Supplier']['PartyLegalEntity']['RegistrationName'];
-                $arrayToReturn['Customer_CUI']  = $arrayParties['Customer']['PartyTaxScheme']['01']['CompanyID'];
-                $arrayToReturn['Customer_Name'] = $arrayParties['Customer']['PartyLegalEntity']['RegistrationName'];
-                unlink($arrayData['strArchivedFileName']);
-                break;
-            case 'XmlFile_Small':
-                $objErrors                      = new \SimpleXMLElement($arrayData['strInvoiceContent']);
-                $arrayToReturn['LoadingIndex']  = $objErrors->attributes()->Index_incarcare->__toString();
-                $arrayToReturn['Size']          = $arrayData['Size'];
-                $arrayToReturn['Supplier_CUI']  = $objErrors->attributes()->Cif_emitent->__toString();
-                $arrayToReturn['Supplier_Name'] = '??????????';
-                $arrayToReturn['Error']         = $objErrors->Error->attributes()->errorMessage->__toString();
-                break;
+        if ($arrayData['Size'] > 1000) {
+            file_put_contents($arrayData['strArchivedFileName'], $arrayData['strInvoiceContent']);
+            $appR                           = new \danielgp\efactura\ClassElectronicInvoiceRead();
+            $arrayElectronicInv             = $appR->readElectronicInvoice($arrayData['strArchivedFileName']);
+            $arrayBasic                     = $arrayElectronicInv['Header']['CommonBasicComponents-2'];
+            $arrayAggregate                 = $arrayElectronicInv['Header']['CommonAggregateComponents-2'];
+            $floatAmounts                   = [
+                'wo_VAT' => (float) $arrayAggregate['LegalMonetaryTotal']['TaxExclusiveAmount']['value'],
+                'TOTAL'  => (float) $arrayAggregate['LegalMonetaryTotal']['TaxInclusiveAmount']['value'],
+            ];
+            $arrayParties                   = [
+                'Customer' => $arrayAggregate['AccountingCustomerParty']['Party'],
+                'Supplier' => $arrayAggregate['AccountingSupplierParty']['Party'],
+            ];
+            $arrayToReturn['LoadingIndex']  = substr($arrayData['Matches'][0][0], 0, -4);
+            $arrayToReturn['Size']          = $arrayData['Size'];
+            $arrayToReturn['Document_No']   = $arrayBasic['ID'];
+            $arrayToReturn['Issue_Date']    = $arrayBasic['IssueDate'];
+            $arrayToReturn['Amount_wo_VAT'] = $floatAmounts['wo_VAT'];
+            $arrayToReturn['Amount_TOTAL']  = $floatAmounts['TOTAL'];
+            $arrayToReturn['Amount_VAT']    = ($floatAmounts['TOTAL'] - $floatAmounts['wo_VAT']);
+            $arrayToReturn['Supplier_CUI']  = $arrayParties['Supplier']['PartyTaxScheme']['01']['CompanyID'];
+            $arrayToReturn['Supplier_Name'] = $arrayParties['Supplier']['PartyLegalEntity']['RegistrationName'];
+            $arrayToReturn['Customer_CUI']  = $arrayParties['Customer']['PartyTaxScheme']['01']['CompanyID'];
+            $arrayToReturn['Customer_Name'] = $arrayParties['Customer']['PartyLegalEntity']['RegistrationName'];
+            unlink($arrayData['strArchivedFileName']);
+        } elseif ($arrayData['Size'] > 0) {
+            $objErrors                      = new \SimpleXMLElement($arrayData['strInvoiceContent']);
+            $arrayToReturn['LoadingIndex']  = $objErrors->attributes()->Index_incarcare->__toString();
+            $arrayToReturn['Size']          = $arrayData['Size'];
+            $arrayToReturn['Supplier_CUI']  = $objErrors->attributes()->Cif_emitent->__toString();
+            $arrayToReturn['Supplier_Name'] = '??????????';
+            $arrayToReturn['Error']         = $objErrors->Error->attributes()->errorMessage->__toString();
         }
         return $arrayToReturn;
     }
